pyproject.toml
[tool.poetry]
name = "dabapush"
version = "0.2.0"
description = ""
authors = ["Philipp Kessling <p.kessling@leibniz-hbi.de>", "Felix Victor Münch <f.muench@leibniz-hbi.de>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/Leibniz-HBI/dabapush"
classifiers = [
		"Development Status :: 2 - Pre-Alpha",
		"Environment :: Console",
		"Intended Audience :: Science/Research",
]
packages = [
	{include = "smo-database"},
	{include = "dabapush"}
]
include = ["config.yml"]

[tool.poetry.dependencies]
python = ">3.8"
click = "^8.0.0"
SQLAlchemy = "^1.4.22"
loguru = "^0.5.3"
pandas = "^1.3.1"
PyYAML = "^5.4.1"
ndjson = "^0.3.1"
ujson = "^5.1.0"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]
pytest = "*"
flake8 = "*"
ipykernel = "^6.9.0"
pdoc3 = "^0.10.0"
pyment = "^0.3.3"
pylint = "^2.13.5"

[tool.poetry.scripts]
dabapush = 'dabapush.main:cli'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
